+++ test/assets/pass/for_loop.lua
@@ -21,3 +21,19 @@ end
 -- Expect: jump=6
 -- Expect: jump=8
 -- Expect: jump=10
+
+function foo(x) return x end
+
+-- Expect: i=1
+-- Expect: i=2
+-- Expect: i=3
+-- Expect: i=4
+-- Expect: i=5
+-- Expect: i=6
+-- Expect: i=7
+-- Expect: i=8
+-- Expect: i=9
+-- Expect: i=10
+for i=foo(1), foo(10), foo(1) do
+	print('i='..i)
+end